+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Added `aggregate()`, `count()`, `sum()`, `average()`, `min()`, and `max()` methods to select queries.
 - Added a connection manager for managing connections to one or more databases.
+- Added `whereInfix()` and `orWhereInfix()` for infix style where conditions.
 
 ## 1.3 - 2016-04-30
 ### Added